Fixes for lightmap + new single-pass terrain renderer
Additive effects color (from things like landing zone lights, explosion effects, etc) was effectively not visible or lighting terrain when using the new single-pass renderer. This was a byproduct of how most lighting calculations have moved to the shader, and the lightmap value (and how additive effect lighting was accumulated) was being calculated differently.

This adjusts the lightmap to now always be 4 channels. When using the new single-pass shaders, the lightmap contains:
1. Additive color from effects (etc) in the .rgb channels
2. The calculated ambient occlusion value in the 4th (.a) channel

The shaders can then handle these two values separately.

(When using the old fallback terrain shaders, the first 3 channels contain the old calculated fully-inclusive-but-not-the-most-accurate illumination value. Only minor changes were required to these shaders to ignore the 4th channel.)

+++ src/display.cpp
@@ -2084,11 +2084,11 @@ void	dealWithLMB()
 			flipVal += 1;
 		}
 
-		console("%s tile %d, %d [%d, %d] continent(l%d, h%d) level %g illum %d ao %g col %x %s %s w=%d s=%d j=%d tile#%d (decal=%s, ground [#%d, size=%.3f], f%d r%d)",
+		console("%s tile %d, %d [%d, %d] continent(l%d, h%d) level %g illum %d ao %d col %x %s %s w=%d s=%d j=%d tile#%d (decal=%s, ground [#%d, size=%.3f], f%d r%d)",
 		        tileIsExplored(psTile) ? "Explored" : "Unexplored",
 		        mouseTileX, mouseTileY, world_coord(mouseTileX), world_coord(mouseTileY),
 		        (int)psTile->limitedContinent, (int)psTile->hoverContinent, psTile->level, (int)psTile->illumination,
-				psTile->ambientOcclusion, psTile->colour.rgba,
+				(int)psTile->ambientOcclusion, psTile->colour.rgba,
 		        aux & AUXBITS_DANGER ? "danger" : "", aux & AUXBITS_THREAT ? "threat" : "",
 		        (int)psTile->watchers[selectedPlayer], (int)psTile->sensors[selectedPlayer], (int)psTile->jammers[selectedPlayer],
 				TileNumber_tile(psTile->texture), (TILE_HAS_DECAL(psTile)) ? "y" : "n",
